Rename MultipleChoice component to Radiobox for clarity, create the bootcamp form.
