+++ frontend/src/components/BootCampForm.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const BootCampForm = () => {
+  return (
+    <form>
+      <div>
+        <strong>Bootcamp Questions</strong> add description...
+      </div>
+    </form>
+  );
+};
+
+export default BootCampForm;
